requirements.txt
certifi==2022.9.24
lxml==4.9.1
numba==0.56.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandas-ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pyzmq==24.0.1
qgrid==1.3.1
requests==2.28.1
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.0
ta==0.10.2
tabulate==0.9.0
tenacity==8.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.1
ujson==5.5.0
urllib3==1.26.12
vectorbt==0.24.2
virtualenv==20.16.6
wcwidth==0.2.5
websocket-client==1.4.1
websockets==9.1
yarl==1.6.3
yfinance==0.1.84
